requirements/base.txt
Django==3.0.3
django-environ==0.4.5

psycopg2-binary==2.8.4

requests==2.23.0
djangorestframework==3.11.0
